asset: move unit test TestGroupKeyIsEqual to group_key_test.go
